Refactor task scheduling system to use a unified scheduler
- Replaced the simple scheduler with a unified task scheduler that integrates both task group definitions and database tasks.
- Updated relevant imports and function calls across the codebase to reflect the new scheduler architecture.
- Enhanced task management by synchronizing database tasks and improving logging for task execution.
- Removed the deprecated simple scheduler module and adjusted tests to ensure compatibility with the new scheduler.
- Added new metrics collection and system maintenance tasks to improve functionality and performance.
